refactor: move some methods to shared, add 2 new methods
